dwebsummit/requirements.txt
django==1.11.13
python-dotenv==0.6.0
pyyaml==3.12
Jinja2==2.10
django-stdimage==2.4.2
django-tinymce==2.7.0
